requirements.txt
# For llm
openai
google-generativeai

# Templates
jinja2

# Others
python-dotenv